+++ app/src/main/java/com/yeongjin/news/data/repository/NewsPagingRepository.kt
@@ -1,7 +1,7 @@
 package com.yeongjin.news.data.repository
 
-import com.yeongjin.news.data.remote.CategoryNewsPagingSource
-import com.yeongjin.news.data.remote.NewsPagingSource
+import com.yeongjin.news.data.remote.paging.CategoryNewsPagingSource
+import com.yeongjin.news.data.remote.paging.NewsPagingSource
 
 interface NewsPagingRepository {
     fun getTopNewsList(country: String): NewsPagingSource
